ai_HalluDomainBench-main/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "halludomainbench"
version = "0.1.0"
description = "HalluDomainBench: a benchmark and platform for LLM domain hallucination security."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiohttp>=3.9.0",
  "dnspython>=2.6.1",
  "requests>=2.31.0",
  "tenacity>=8.2.0",
  "tldextract>=5.1.2",
]

[project.scripts]
halludomainbench = "halludomainbench.cli:main_entry"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
